# server/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API proyectoUTM",
        "description": "<html><body>Esta es la documentación de la API para el sistema de Investigación Docencia.<p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los servicios proporcionados.</p><br><p><b>Autenticación y token de API</b></p><p>La Universidad Tecnológica de la Mixteca (UTM) proporciona el api-token. Una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/institutos/eliminar/{idInstituto}": {
            "delete": {
                "summary": "Eliminar un instituto",
                "operationId": "eliminar",
                "parameters": [{
                        "description": "apiKey",
                        "type": "string",
                        "in": "header",
                        "name": "Authorization",
                        "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                    },
                    {
                        "name": "idInstituto",
                        "in": "path",
                        "description": "id del Instituto a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/institutos/actualizar/{idInstituto}": {
            "put": {
                "summary": "Modificando un instituto",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "description": "apiKey",
                        "type": "string",
                        "in": "header",
                        "name": "Authorization",
                        "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                    }, {
                        "name": "idInstituto",
                        "in": "path",
                        "description": "id del Instituto a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Instituto",
                        "in": "body",
                        "description": "Instituto a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoInstituto"
                        }
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/carrerass/eliminar/{idCarrera}": {
            "delete": {
                "summary": "Eliminar una carrera",
                "operationId": "eliminar",
                "parameters": [{
                    "description": "apiKey",
                    "type": "string",
                    "in": "header",
                    "name": "Authorization",
                    "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                }, {
                    "name": "idCarrera",
                    "in": "path",
                    "description": "id de la Carrera a eliminar",
                    "required": true,
                    "default": 14
                }],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/carreras/actualizar/{idCarrera}": {
            "put": {
                "summary": "Modificando una carrera",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "description": "apiKey",
                        "type": "string",
                        "in": "header",
                        "name": "Authorization",
                        "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                    }, {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "id de la Carrera a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Carrera",
                        "in": "body",
                        "description": "Carrera a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejaCarrera"
                        }
                    }
                ],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/profesores/eliminar/{idProfesor}": {
            "delete": {
                "summary": "Eliminar un profesor",
                "operationId": "eliminar",
                "parameters": [{
                    "description": "apiKey",
                    "type": "string",
                    "in": "header",
                    "name": "Authorization",
                    "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                }, {
                    "name": "idProfesor",
                    "in": "path",
                    "description": "id del profesor a eliminar",
                    "required": true,
                    "default": 14
                }],
                "tags": [
                    "Profesores"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/profesores/actualizar/{idProfesor}": {
            "put": {
                "summary": "Modificando una Profesor",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "description": "apiKey",
                        "type": "string",
                        "in": "header",
                        "name": "Authorization",
                        "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                    }, {
                        "name": "idProfesor",
                        "in": "path",
                        "description": "id de la Profesor a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Profesor",
                        "in": "body",
                        "description": "Profesor a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoProfesor"
                        }
                    }
                ],
                "tags": [
                    "Profesores"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/articulo/eliminar/{idArticulos}": {
            "delete": {
                "summary": "Eliminar un Articulos",
                "operationId": "eliminar",
                "parameters": [{
                    "description": "apiKey",
                    "type": "string",
                    "in": "header",
                    "name": "Authorization",
                    "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                }, {
                    "name": "idArticulos",
                    "in": "path",
                    "description": "Identificador del Articulos a eliminar",
                    "required": true,
                    "default": 14
                }],
                "tags": [
                    "Articulos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/api/articulo/actualizar/{idArticulos}": {
            "put": {
                "summary": "Modificando una Articulos",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                        "description": "apiKey",
                        "type": "string",
                        "in": "header",
                        "name": "Authorization",
                        "default": "Bearer eyJhbGciOiJIUzI1NiJ9.cmhldmU4NTVAZ21haWwuY29t.k1Elo10OhVf_UJKHERDIZSdlPUQZ1RHDjbp_bIqOgiA"
                    }, {
                        "name": "idArticulos",
                        "in": "path",
                        "description": "id de la Articulos a actualizar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Articulos",
                        "in": "body",
                        "description": "Articulos a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoArticulo"
                        }
                    }
                ],
                "tags": [
                    "Articulos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ViejoInstituto": {
            "type": "object",
            "required": [
                "codigo",
                "nombre"
            ],
            "properties": {
                "codigo": {
                    "type": "string",
                    "default": "111"
                },
                "nombre": {
                    "type": "string",
                    "default": "Instituto patito"
                }
            }
        },
...
        "ViejaCarrera": {
            "type": "object",
            "required": [
                "codigo",
                "nombre",
                "idInstituto"
            ],
            "properties": {
                "codigoCarrera": {
                    "type": "string",
                    "default": "011"
                },
                "nombreCarrera": {
                    "type": "string",
                    "default": "Carrera patito"
                },
                "idInstituto": {
                    "type": "number",
                    "default": "2"
                }
            }
        },
...
        "ViejoProfesor": {
            "type": "object",
            "required": [
                "nombresP",
                "apellidoP",
                "apellidoM",
                "correoProfesor",
                "password",
                "nivel",
                "idCarrera",
                "grado",
                "idTipoProfesor",
                "nombreApa",
                "idInstituto"
            ],
            "properties": {
                "nombresP": {
                    "type": "string",
                    "default": "Nombre Patito"
                },
                "apellidoP": {
                    "type": "string",
                    "default": "Apellido Paterno Patito"
                },
                "apellidoM": {
                    "type": "string",
                    "default": "Apellido Materno Patito"
                },
                "correoProfesor": {
                    "type": "string",
                    "default": "Correo Patito"
                },
                "password": {
                    "type": "string",
                    "default": "Contraseña Patito"
                },
                "nivel": {
                    "type": "number",
                    "default": "2"
                },
                "idCarrera": {
                    "type": "number",
                    "default": "2"
                },
                "grado": {
                    "type": "string",
                    "default": "Grado Patito"
                },
                "idTipoProfesor": {
                    "type": "number",
                    "default": "2"
                },
                "nombreApa": {
                    "type": "string",
                    "default": "Nombre Apa Patito"
                },
                "idInstituto": {
                    "type": "number",
                    "default": "2"
                }
            }
        },
...
        "ViejoArticulo": {
            "type": "object",
            "required": [
                "tipCLR",
                "titulo",
                "nombreCLR",
                "estado",
                "fechaEdicion",
                "tipoNI",
                "issnisbn",
                "volumen",
                "paginas",
                "anyo",
                "doi",
                "comprobante",
                "indexa",
                "issue",
                "editores",
                "cuidad",
                "pais",
                "editorial"
            ],
            "properties": {
                "tipCLR": {
                    "type": "string",
                    "default": "Revista"
                },
                "titulo": {
                    "type": "string",
                    "default": "Patito"
                },
                "nombreCLR": {
                    "type": "string",
                    "default": "Revista Patito"
                },
                "estado": {
                    "type": "string",
                    "default": "Publicado"
                },
                "fechaEdicion": {
                    "type": "string",
                    "default": "2022-01-25"
                },
                "tipoNI": {
                    "type": "string",
                    "default": "Nacional"
                },
                "issnisbn": {
                    "type": "string",
                    "default": "Patito123"
                },
                "volumen": {
                    "type": "string",
                    "default": "1"
                },
                "paginas": {
                    "type": "string",
                    "default": "15-20"
                },
                "anyo": {
                    "type": "string",
                    "default": "2022"
                },
                "doi": {
                    "type": "string",
                    "default": "http://programacion.com"
                },
                "comprobante": {
                    "type": "string",
                    "default": ""
                },
                "indexa": {
                    "type": "string",
                    "default": "1"
                },
                "issue": {
                    "type": "string",
                    "default": ""
                },
                "editores": {
                    "type": "string",
                    "default": "Tres Patitos"
                },
                "cuidad": {
                    "type": "string",
                    "default": "Huajuapan"
                },
                "pais": {
                    "type": "string",
                    "default": "Mexico"
                },
                "editorial": {
                    "type": "string",
                    "default": "Patito"
                }
            }
        }
    }
}